requirements.txt
opencv-python>=4.8.0
pillow>=9.0.0
numpy>=1.21.0
scikit-image>=0.19.0
